small fixes

+++ src/main/java/com/github/dirtpowered/betatorelease/proxy/translator/clientbound/ServerEntityPositionRotationTranslator.java
@@ -15,8 +15,8 @@ public class ServerEntityPositionRotationTranslator implements ModernToBetaHandl
         int y = Utils.toAbsolutePos(packet.getMovementY());
         int z = Utils.toAbsolutePos(packet.getMovementZ());
 
-        byte yaw = (byte) Utils.toAbsoluteRotation(packet.getYaw());
-        byte pitch = (byte) Utils.toAbsoluteRotation(packet.getPitch());
+        int yaw = Utils.toAbsoluteRotation(packet.getYaw());
+        int pitch = Utils.toAbsoluteRotation(packet.getPitch());
 
         betaSession.sendPacket(new V1_7_3EntityMoveLookPacketData(entityId, x, y, z, yaw, pitch));
     }
