+++ src/main/java/com/github/dirtpowered/betatorelease/proxy/translator/clientbound/ServerSpawnMobTranslator.java
@@ -27,8 +27,8 @@ public class ServerSpawnMobTranslator implements ModernToBetaHandler<ServerSpawn
         int x = Utils.toAbsolutePos(packet.getX());
         int y = Utils.toAbsolutePos(packet.getY());
         int z = Utils.toAbsolutePos(packet.getZ());
-        byte yaw = (byte) Utils.toAbsoluteRotation((int) packet.getYaw());
-        byte pitch = (byte) Utils.toAbsoluteRotation((int) packet.getPitch());
+        byte yaw = (byte) Utils.toAbsoluteRotation(packet.getYaw());
+        byte pitch = (byte) Utils.toAbsoluteRotation(packet.getPitch());
 
         if (mobType == MobType.VILLAGER) {
             betaSession.sendPacket(new V1_7_3NamedEntitySpawnPacketData(entityId, NAME, x, y, z, yaw, pitch, 264));
